refactor(agent): replace core middleware with policy engine
