+++ packages/agent/src/core/index.ts
@@ -20,25 +20,13 @@ export type {
   RuntimeAgentInstance,
   RuntimeInstanceStatus,
 } from "./runtime-context";
-export { MiddlewareEngine } from "./middleware";
-export type {
-  MiddlewareContext,
-  MiddlewareFn,
-  MiddlewareRegistration,
-  MiddlewareDecision,
-  MiddlewareAuditConfig,
-  MiddlewareEngineConfig,
-  MiddlewareEngineInstance,
-} from "./middleware";
 export { PolicyEngine } from "./policy";
 export type {
-  PolicyAuditConfig,
   PolicyContext,
-  PolicyDecision,
-  PolicyEngineConfig,
-  PolicyEngineInstance,
   PolicyFn,
   PolicyRegistration,
-  PolicySystemPromptVerdict,
-  PolicyVerdict,
+  PolicyDecision,
+  PolicyAuditConfig,
+  PolicyEngineConfig,
+  PolicyEngineInstance,
 } from "./policy";
