--- packages/agent/src/core/middleware/builtin/index.ts
@@ -1,10 +1,0 @@
-export { createBudgetReassuranceMiddleware, createBudgetWarningMiddleware } from "./budget";
-export { createCompactionMiddleware } from "./compaction";
-export { createPostToolMiddleware } from "./post-tool";
-export type { PostToolEnricher } from "./post-tool";
-export { createPostTurnMiddleware } from "./post-turn";
-export type { PostTurnHandler } from "./post-turn";
-export { createIdleNudgeMiddleware } from "./idle-nudge";
-export type { IdleNudgeConfig } from "./idle-nudge";
-export { createToolGuardMiddleware } from "./tool-guard";
-export type { ToolGuardMiddlewareConfig } from "./tool-guard";
