--- packages/agent/src/core/middleware/index.ts
@@ -1,10 +1,0 @@
-export type { MiddlewareContext, MiddlewareFn, MiddlewareRegistration } from "./types";
-export { MiddlewareEngine } from "./engine";
-export { fromConfig } from "./compat";
-export type {
-  MiddlewareDecision,
-  MiddlewareAuditConfig,
-  MiddlewareEngineConfig,
-  MiddlewareEngineInstance,
-} from "./engine";
-export * from "./builtin";
