+++ packages/agent/src/core/policy/index.ts
@@ -1,15 +1,9 @@
+export type { PolicyContext, PolicyFn, PolicyRegistration } from "./types";
 export { PolicyEngine } from "./engine";
 export type {
-  PolicyAuditConfig,
   PolicyDecision,
+  PolicyAuditConfig,
   PolicyEngineConfig,
-} from "./engine";
-export type {
-  PolicyContext,
   PolicyEngineInstance,
-  PolicyFn,
-  PolicyRegistration,
-  PolicySystemPromptVerdict,
-  PolicyVerdict,
-} from "./types";
+} from "./engine";
 export * from "./builtin";
