+++ packages/agent/src/index.ts
@@ -1,4 +1,4 @@
-// Agent package public API
+// Agent package public API — ChatAgent only
 export { ChatAgent } from "./core/chat-agent";
 export type { ChatAgentInstance } from "./core/chat-agent";
 export type {
@@ -10,12 +10,7 @@ export type {
   AgentBudget,
   TokenUsage,
   Sink,
-  StepGuardVerdict,
-  StepGuardContext,
   AgentEventEmitter,
-  ExecutionHooks,
-  HookContext,
-  HookVerdict,
 } from "./core/types";
 export {
   createAgentRuntimeContext,
@@ -29,27 +24,15 @@ export type {
   RuntimeAgentInstance,
   RuntimeInstanceStatus,
 } from "./core/runtime-context";
-export { MiddlewareEngine } from "./core/middleware";
-export type {
-  MiddlewareContext,
-  MiddlewareFn,
-  MiddlewareRegistration,
-  MiddlewareDecision,
-  MiddlewareAuditConfig,
-  MiddlewareEngineConfig,
-  MiddlewareEngineInstance,
-} from "./core/middleware";
 export { PolicyEngine } from "./core/policy";
 export type {
-  PolicyAuditConfig,
   PolicyContext,
-  PolicyDecision,
-  PolicyEngineConfig,
-  PolicyEngineInstance,
   PolicyFn,
   PolicyRegistration,
-  PolicySystemPromptVerdict,
-  PolicyVerdict,
+  PolicyDecision,
+  PolicyAuditConfig,
+  PolicyEngineConfig,
+  PolicyEngineInstance,
 } from "./core/policy";
 export { AgentRegistry } from "./runtime/index";
 export { SubagentTool } from "./runtime/index";
@@ -59,9 +42,9 @@ export type { BackgroundOutputToolOptions, BackgroundCancelToolOptions } from ".
 export { McpClient } from "./runtime/mcp/index";
 export type { McpServerConfig } from "./runtime/mcp/index";
 export {
-  createBudgetReassuranceMiddleware,
-  createBudgetWarningMiddleware,
-} from "./core/middleware/builtin/budget";
-export { createIdleNudgeMiddleware } from "./core/middleware/builtin/idle-nudge";
-export { createToolGuardMiddleware } from "./core/middleware/builtin/tool-guard";
+  createBudgetReassurancePolicy,
+  createBudgetWarningPolicy,
+} from "./core/policy/builtin/budget";
+export { createIdleNudgePolicy } from "./core/policy/builtin/idle-nudge";
+export { createToolPermissionPolicy } from "./core/policy/builtin/tool-guard";
 export { InMemoryCompactor } from "./core/execution/compaction";
