--- packages/agent/test/core/step-guard.test.ts
@@ -1,164 +1,0 @@
-import { beforeAll, beforeEach, describe, expect, it, mock } from "bun:test";
-import type { Sink } from "@openomni/protocol";
-import type { StepGuardContext } from "../../src/core/types";
-import {
-  createStopOutcome,
-  createMockLlmConfig,
-  mockProviderData,
-  mockProviderModel,
-  type MockLlmFn,
-} from "../helpers/mock-llm";
-
-let mockRunFn: MockLlmFn = async () => createStopOutcome();
-
-const mockLlm = createMockLlmConfig({
-  getModels: mock(async () => mockProviderData),
-  fromModelsDevModel: mock(() => mockProviderModel),
-  run: (input, sink: Sink) => mockRunFn(input, sink),
-});
-
-let ChatAgent: typeof import("../../src/core/chat-agent").ChatAgent;
-
-beforeAll(async () => {
-  ({ ChatAgent } = await import("../../src/core/chat-agent"));
-});
-
-beforeEach(() => {
-  mockRunFn = async () => createStopOutcome();
-});
-
-describe("StepGuard (run path)", () => {
-  it("runs normally when no stepGuard configured", async () => {
-    const agent = ChatAgent.create({
-      model: { provider: "anthropic", id: "claude-3-haiku-20240307" },
-      llm: mockLlm,
-    });
-    const result = await agent.run({
-      messages: [{ role: "user", content: "hello" }],
-    });
-    expect(result.finishReason).toBe("stop");
-    expect(result.guardAborted).toBeUndefined();
-  });
-
-  it("stops normally when guard returns continue", async () => {
-    const agent = ChatAgent.create({
-      model: { provider: "anthropic", id: "claude-3-haiku-20240307" },
-      llm: mockLlm,
-      stepGuard: async () => ({ action: "continue" }),
-    });
-    const result = await agent.run({
-      messages: [{ role: "user", content: "hello" }],
-    });
-    expect(result.finishReason).toBe("stop");
-    expect(result.guardAborted).toBeUndefined();
-  });
-
-  it("injects message and continues when guard returns inject", async () => {
-    let callCount = 0;
-    mockRunFn = async () => {
-      callCount++;
-      return createStopOutcome();
-    };
-
-    let guardCallCount = 0;
-    const agent = ChatAgent.create({
-      model: { provider: "anthropic", id: "claude-3-haiku-20240307" },
-      llm: mockLlm,
-      stepGuard: async () => {
-        guardCallCount++;
-        if (guardCallCount === 1) {
-          return {
-            action: "inject",
-            message: "Please verify your work.",
-            reason: "verify-work",
-            policyId: "test.step-guard",
-          };
-        }
-        return { action: "continue" };
-      },
-    });
-
-    const result = await agent.run({
-      messages: [{ role: "user", content: "do task" }],
-    });
-    expect(result.finishReason).toBe("stop");
-    expect(result.guardAborted).toBeUndefined();
-    expect(callCount).toBe(2);
-    expect(guardCallCount).toBe(2);
-  });
-
-  it("returns with guardAborted when guard aborts", async () => {
-    const agent = ChatAgent.create({
-      model: { provider: "anthropic", id: "claude-3-haiku-20240307" },
-      llm: mockLlm,
-      stepGuard: async () => ({ action: "abort", reason: "too long" }),
-    });
-    const result = await agent.run({
-      messages: [{ role: "user", content: "hello" }],
-    });
-    expect(result.finishReason).toBe("stop");
-    expect(result.guardAborted).toBe(true);
-  });
-
-  it("provides correct context to guard", async () => {
-    let capturedContext: StepGuardContext | null = null;
-    const agent = ChatAgent.create({
-      model: { provider: "anthropic", id: "claude-3-haiku-20240307" },
-      llm: mockLlm,
-      stepGuard: async (_step, ctx) => {
-        capturedContext = ctx;
-        return { action: "continue" };
-      },
-    });
-    await agent.run({
-      messages: [{ role: "user", content: "hello" }],
-    });
-    expect(capturedContext).not.toBeNull();
-    if (!capturedContext) {
-      throw new Error("Expected step guard context");
-    }
-    expect(capturedContext.isCompletion).toBe(true);
-    expect(capturedContext.continuationCount).toBe(0);
-    expect(capturedContext.elapsedMs).toBeGreaterThanOrEqual(0);
-  });
-
-  it("increments continuationCount on each inject", async () => {
-    mockRunFn = async () => createStopOutcome();
-
-    const capturedCounts: number[] = [];
-    const agent = ChatAgent.create({
-      model: { provider: "anthropic", id: "claude-3-haiku-20240307" },
-      llm: mockLlm,
-      stepGuard: async (_step, ctx) => {
-        capturedCounts.push(ctx.continuationCount);
-        if (ctx.continuationCount < 2) {
-          return {
-            action: "inject",
-            message: "keep going",
-            reason: "continue-work",
-            policyId: "test.step-guard",
-          };
-        }
-        return { action: "continue" };
-      },
-    });
-    await agent.run({
-      messages: [{ role: "user", content: "task" }],
-    });
-    expect(capturedCounts).toEqual([0, 1, 2]);
-  });
-
-  it("continues normally when guard throws (fail-open)", async () => {
-    // Guard errors are now handled by middleware engine with fail-open policy:
-    // the error is logged and execution continues rather than propagating.
-    const agent = ChatAgent.create({
-      model: { provider: "anthropic", id: "claude-3-haiku-20240307" },
-      llm: mockLlm,
-      stepGuard: async () => {
-        throw new Error("guard error");
-      },
-    });
-    const result = await agent.run({ messages: [{ role: "user", content: "hello" }] });
-    expect(result.finishReason).toBe("stop");
-  });
-});
